Feat/voltage-sequence (#20)
Working VoltageSequence for GateSet and VirtualGateSet. 
Default behaviour for VoltageSequence is now to remember it's last point and update virtual gates as required instead of setting unspecified ones to 0, can be turned off when instantiating VoltageSequence.
Also some enforced setting voltages to 0 occurs when applying the compensation pulse, this is to ensure calculations are performed at 0 offset.